+++ src/Symfony/Component/Form/Extension/Validator/ValidatorExtension.php
@@ -1,0 +1,38 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien.potencier@symfony-project.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Form\Extension\Validator;
+
+use Symfony\Component\Form\Extension\Validator\Type;
+use Symfony\Component\Form\AbstractExtension;
+use Symfony\Component\Validator\ValidatorInterface;
+
+class ValidatorExtension extends AbstractExtension
+{
+    private $validator;
+
+    public function __construct(ValidatorInterface $validator)
+    {
+        $this->validator = $validator;
+    }
+
+    protected function loadTypeExtensions()
+    {
+        return array(
+            new Type\FieldTypeValidatorExtension($this->validator),
+        );
+    }
+
+    public function loadTypeGuesser()
+    {
+        return new ValidatorTypeGuesser($this->validator->getMetadataFactory());
+    }
+}
