+++ src/Symfony/Component/Form/Extension/Validator/ValidatorTypeGuesser.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Form\Extension\Core;
+namespace Symfony\Component\Form\Extension\Validator;
 
 use Symfony\Component\Form\FormTypeGuesserInterface;
 use Symfony\Component\Form\Guess\Guess;
@@ -17,7 +17,7 @@ use Symfony\Component\Form\Guess\TypeGuess;
 use Symfony\Component\Form\Guess\ValueGuess;
 use Symfony\Component\Validator\Mapping\ClassMetadataFactoryInterface;
 
-class CoreTypeGuesser implements FormTypeGuesserInterface
+class ValidatorTypeGuesser implements FormTypeGuesserInterface
 {
     private $metadataFactory;
